star-edge/image/edge/app/requirements.txt
#websocket-client==0.56.0
#pyserial==3.4

websocket-client
pyserial